backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
python-pptx==0.6.23
pdf2image==1.16.3
Pillow==10.1.0

# LLM
openai==1.12.0

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
